Create MutualSubstitution.cpp
A mutual letter substitution is a pair of letters subFirst, subSecond, and denotes an operation where letter subSecond is substituted in the place of every occurrence of letter subFirst, and vice versa. Given a string of n lowercase letters, and m mutual substitution pair of letters: subFirst and subSecond. Goal is the perform all the substitutions and return the resulting string.